+++ contrastsecurity/datadog_checks/contrastsecurity/__init__.py
@@ -1,4 +1,0 @@
-from .__about__ import __version__
-from .contrastsecurity import ContrastsecurityCheck
-
-__all__ = ['__version__', 'ContrastsecurityCheck']