+++ lighthouse/tests/test_lighthouse.py
@@ -1,9 +1,10 @@
 import os
-import pytest
 
+import pytest
 from mock import MagicMock
-from datadog_checks.lighthouse import LighthouseCheck
+
 from datadog_checks.errors import CheckException
+from datadog_checks.lighthouse import LighthouseCheck
 
 HERE = os.path.dirname(os.path.abspath(__file__))