+++ logstash/datadog_checks/logstash/logstash.py
@@ -179,13 +179,15 @@ class LogstashCheck(AgentCheck):
             version = data['version']
         except Exception as e:
             self.warning(
-                "Error while trying to get Logstash version "
-                "from %s %s. Defaulting to version %s." % (config.url, str(e), self.DEFAULT_VERSION)
+                "Error while trying to get Logstash version from %s %s. Defaulting to version %s.",
+                config.url,
+                e,
+                self.DEFAULT_VERSION,
             )
             version = self.DEFAULT_VERSION
 
         self.service_metadata('version', version)
-        self.log.debug("Logstash version is %s" % version)
+        self.log.debug("Logstash version is %s", version)
         return version
 
     def check(self, instance):