+++ traefik/datadog_checks/traefik/traefik.py
@@ -34,12 +34,12 @@ class TraefikCheck(AgentCheck):
                         self.gauge('traefik.total_status_code_count', count, ['status_code:{}'.format(status_code)])
 
                 else:
-                    self.log.warn('Field total_status_code_count not found in response.')
+                    self.log.warning('Field total_status_code_count not found in response.')
 
                 if 'total_count' in payload:
                     self.gauge('traefik.total_count', payload['total_count'])
                 else:
-                    self.log.warn('Field total_count not found in response.')
+                    self.log.warning('Field total_count not found in response.')
 
             else:
                 self.service_check(